LayoutTests/media/modern-media-controls/aspect-ratio-button/aspect-ratio-button.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>AspectRatioButton</code> class.");

const scaleToFillButton = new AspectRatioButton({ layoutTraits: LayoutTraits.macOS });

shouldBeEqualToString("scaleToFillButton.element.localName", "button");
shouldBeTrue('scaleToFillButton.element.classList.contains("icon")');
shouldBeTrue('scaleToFillButton.element.classList.contains("aspect-ratio")');
shouldBeEqualToString("scaleToFillButton.iconName", `${Icons.ScaleToFill}`);

const scaleToFitButton = new AspectRatioButton({ layoutTraits: LayoutTraits.macOS });
scaleToFitButton.scalesToFill = true;
shouldBeEqualToString("scaleToFitButton.iconName", `${Icons.ScaleToFit}`);

scheduler.frameDidFire = function()
{
    if ([scaleToFillButton, scaleToFitButton].some(button => button.width == 0))
        return;

    shouldBeTrue('scaleToFillButton.element.style.webkitMaskImage.includes("macOS/scale-to-fill@")');
    shouldBeTrue('scaleToFitButton.element.style.webkitMaskImage.includes("macOS/scale-to-fit@")');

    finishJSTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
